Refactor User Modal to support dynamic roles
Refactor NewStaffModal to a generic NewUserModal component - Changed modal to accommodate dynamic user roles (e.g., staff, client) - Removed specific NewStaffModal component in favor of NewUserModal - Updated UserTable to accept and display specific roles

Release-Note: Add NewUserModal to handle dynamic role creation for staff and clients
